requirements.txt
fastapi
uvicorn
httpx
python-multipart
python-dotenv
grpcio
grpcio-tools
